refactor(core): move domain types out of tui
