+++ src/core/mod.rs
@@ -1,11 +1,14 @@
 pub mod app;
 pub mod auth;
+pub mod banner;
 pub mod config;
 #[cfg(feature = "cover-art")]
 pub mod cover_theme;
 pub mod first_run;
 pub mod format;
-pub mod layout;
+pub mod geometry;
+pub mod input;
+pub mod limits;
 pub mod migrations;
 pub mod pagination;
 pub mod paths;
@@ -17,4 +20,5 @@ pub mod source;
 pub mod state;
 #[cfg(test)]
 pub mod test_helpers;
+pub mod theme;
 pub mod user_config;
