+++ src/tui/mod.rs
@@ -1,7 +1,8 @@
-pub mod banner;
 #[cfg(feature = "cover-art")]
 pub mod cover_art;
 pub mod event;
 pub mod handlers;
+pub mod layout;
 pub mod runner;
+pub mod theme;
 pub mod ui;
